Forego mega_clusters and fix Gaussian read energy
For generating clusters, it used to be that one would make a giant
cluster and complete radially. This is still the case but now the
dimensions of the giant cluster are automatically detected. Previously
when reading gaussian output the tag would be 'Done'. This has been
updated to 'SCF Done' to avoid confusion with the lines that say
'EDone=F'