Did a lot of work on grid fitting. Not done
Supergrid works now and we can visualise it. However we won't ever need
the grid sorted and fitting on grid points for the fitting procedure so
I have separated the sorting from the fractionalisation part of the
code. Consider carefully if the mod 1 part needs to go back in the
fractionalisation